notes-to-notion/backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.4.2
httpx==0.25.0
pytesseract==0.3.10
Pillow==10.0.1
pdf2image==1.16.3
requests==2.31.0